Move notificators to appropriate package and rename.
