+++ src/org/traccar/notificators/NotificatorWeb.java
@@ -14,21 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar.notification;
+package org.traccar.notificators;
 
-import org.traccar.helper.Log;
+import org.traccar.Context;
 import org.traccar.model.Event;
 import org.traccar.model.Position;
 
-public final class NotificationNull extends Notificator {
-
-    @Override
-    public void sendAsync(long userId, Event event, Position position) {
-        Log.warning("You are using null notificatior, please check your configuration, notification not sent");
-    }
+public final class NotificatorWeb extends Notificator {
 
     @Override
     public void sendSync(long userId, Event event, Position position) {
-        Log.warning("You are using null notificatior, please check your configuration, notification not sent");
+        Context.getConnectionManager().updateEvent(userId, event);
     }
+
 }
